# src/api-clients/backend-openapi-v1/models/DealPatchDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Name } from './Name';
export type DealPatchDTO = {
  name?: Name;
  client?: Name;
};