# src/api-clients/backend-openapi-v1/models/DealsResponseDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { DealResponseDTO } from './DealResponseDTO';
export type DealsResponseDTO = {
  deals: Array<DealResponseDTO>;
  total: number;
};