# src/api-clients/backend-openapi-v1/models/Error.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Error = {
  name: string;
  message: string;
  stack?: string;
};